feat: remuve interceptor
