+++ src/app/app.component.ts
@@ -1,6 +1,7 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { HeaderComponent } from './components/header.component';
+import { ObtainAccessTokenService } from './core/services/api/obtain-access-token';
 
 @Component({
   selector: 'app-root',
@@ -9,4 +10,10 @@ import { HeaderComponent } from './components/header.component';
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
 })
-export class AppComponent {}
+export class AppComponent implements OnInit {
+  constructor(private authService: ObtainAccessTokenService) {}
+
+  ngOnInit(): void {
+    this.authService.configureAccessToken();
+  }
+}
