+++ src/app/environment/environment.ts
@@ -1,0 +1,8 @@
+export const environment = {
+  host: 'https://api.europe-west1.gcp.commercetools.com',
+  auth_url: 'https://auth.europe-west1.gcp.commercetools.com',
+  project_key: 'ecommerce-application-rsschool-1905',
+  client_id: 'X3-sG51LSAsgBux8z_ZPQ4D-',
+  client_secret: 'OMxGBGWWke-g5B-ZfqycNpQpnNIcbmiF',
+  scope: 'manage_project:ecommerce-application-rsschool-1905',
+};
